Dev - fixed comment section and started developing notification system

+++ components/commentSection/CommentSection.tsx
@@ -16,11 +16,21 @@ import Image from 'next/image'
 import { useSession } from 'next-auth/react'
 import { toast } from 'react-toastify'
 import RepliesSection from './RepliesSection'
-import { timeAgo } from '@/utils/functions'
+import { sendCommentNotification, timeAgo } from '@/utils/functions'
 
-type Props = { postId: string }
+type Props = {
+  postId: string
+  postSlug: string
+  postAuthor: string
+  postTitle: string
+}
 
-export default function CommentSection({ postId }: Props) {
+export default function CommentSection({
+  postId,
+  postSlug,
+  postAuthor,
+  postTitle,
+}: Props) {
   const { data: session, status } = useSession()
   const [currentComment, setCurrentComment] = useState<string>('')
   const [hideComments, setHideComments] = useState<boolean>(true)
@@ -65,6 +75,20 @@ export default function CommentSection({ postId }: Props) {
   useEffect(() => {
     if (comments) initializeReplies()
   }, [comments]) // eslint-disable-line react-hooks/exhaustive-deps
+
+  useEffect(() => {
+    function reset() {
+      SetShowId('')
+    }
+    if (showId) {
+      document.body.addEventListener('click', reset)
+    }
+
+    return () => {
+      document.body.removeEventListener('click', reset)
+    }
+  }, [showId])
+
   const {
     commentSectionContainer,
     mainCommentSection,
@@ -93,6 +117,7 @@ export default function CommentSection({ postId }: Props) {
     opened,
     replyContainer,
     isAuthor,
+    aboveCommentContent,
   } = styles
   async function handleSendComment() {
     if (!session) {
@@ -107,11 +132,22 @@ export default function CommentSection({ postId }: Props) {
         toast.error('something went wrong! Please try again later.', {
           toastId: 'error_add_cmt',
         })
+        setPosting(false)
       } else {
         setCurrentComment('')
         await initializeComments()
+        setPosting(false)
+
+        const actorId = session?.user.id
+        const actor = session?.user.name
+        await sendCommentNotification(
+          actor,
+          actorId,
+          postAuthor,
+          postTitle,
+          postSlug
+        )
       }
-      setPosting(false)
     }
   }
   function handleDropdown(id: string) {
@@ -134,6 +170,8 @@ export default function CommentSection({ postId }: Props) {
   }
   async function handleDelete(id: string) {
     SetShowId('')
+    const sure = confirm('Are you sure you want to delete this comment ?')
+    if (!sure) return
     const repliesDeleted = await deleteCommentReplies(id)
     if (!repliesDeleted) {
       toast.error('something went wrong! Please try again later.', {
@@ -206,33 +244,72 @@ export default function CommentSection({ postId }: Props) {
           {comments &&
             comments.map((comment) => (
               <div key={comment.id} className={commentContainer}>
-                <div className={readerContainer}>
-                  <div className={readerAvatar}>
-                    <Image
-                      src={comment.reader.photo.url}
-                      width={24}
-                      height={24}
-                      alt={comment.reader.name}
-                      style={{ borderRadius: '50%' }}
-                    />
+                <div className={aboveCommentContent}>
+                  <div className={readerContainer}>
+                    <div className={readerAvatar}>
+                      <Image
+                        src={comment.reader.photo.url}
+                        width={24}
+                        height={24}
+                        alt={comment.reader.name}
+                        style={{ borderRadius: '50%' }}
+                      />
+                    </div>
+                    <div
+                      className={`${readerName} ${
+                        comment.reader.id === session?.user.id
+                          ? myComment
+                          : null
+                      }`}
+                    >
+                      {comment.reader.name}
+                    </div>
+                    {comment.reader.isAuthor && (
+                      <div className={isAuthor}>Author</div>
+                    )}
                   </div>
-                  <div
-                    className={`${readerName} ${
-                      comment.reader.id === session?.user.id ? myComment : null
-                    }`}
-                  >
-                    {comment.reader.name}
+                  <div>
+                    {comment.reader.id === session?.user.id && (
+                      <div className={dropdown}>
+                        <button
+                          disabled={editing !== ''}
+                          onClick={() => handleDropdown(comment.id)}
+                        >
+                          <span className={menuDot}></span>
+                          <span className={menuDot}></span>
+                          <span className={menuDot}></span>
+                        </button>
+                        <div
+                          className={`${dropdownContent} ${
+                            showId && showId === comment.id && show
+                          }`}
+                        >
+                          <div
+                            onClick={() => {
+                              if (comment.reader.id !== session?.user.id) return
+                              setEditing(comment.id)
+                              setCurrentEditingComment(comment.comment)
+                              SetShowId('')
+                            }}
+                          >
+                            Edit
+                          </div>
+                          <div onClick={() => handleDelete(comment.id)}>
+                            Delete
+                          </div>
+                        </div>
+                      </div>
+                    )}
                   </div>
-                  {comment.reader.isAuthor && (
-                    <div className={isAuthor}>Author</div>
-                  )}
                 </div>
+
                 <div className={commentContentContainer}>
                   {editing === comment.id ? (
                     <div className={commentEditContainer}>
                       <textarea
                         rows={3}
                         value={currentEditingComment}
+                        autoFocus
                         onChange={(e) =>
                           setCurrentEditingComment(e.target.value)
                         }
@@ -272,41 +349,13 @@ export default function CommentSection({ postId }: Props) {
                     <div className={edited}>Edited</div>
                   )}
                   <div className={age}>{timeAgo(comment.createdAt)}</div>
-                  {comment.reader.id === session?.user.id && (
-                    <div className={dropdown}>
-                      <button
-                        disabled={editing !== ''}
-                        onClick={() => handleDropdown(comment.id)}
-                      >
-                        <span className={menuDot}></span>
-                        <span className={menuDot}></span>
-                        <span className={menuDot}></span>
-                      </button>
-                      <div
-                        className={`${dropdownContent} ${
-                          showId && showId === comment.id && show
-                        }`}
-                      >
-                        <div
-                          onClick={() => {
-                            if (comment.reader.id !== session?.user.id) return
-                            setEditing(comment.id)
-                            setCurrentEditingComment(comment.comment)
-                            SetShowId('')
-                          }}
-                        >
-                          Edit
-                        </div>
-                        <div onClick={() => handleDelete(comment.id)}>
-                          Delete
-                        </div>
-                      </div>
-                    </div>
-                  )}
                 </div>
                 <RepliesSection
                   commentId={comment.id}
                   postId={postId}
+                  postSlug={postSlug}
+                  postAuthor={postAuthor}
+                  postTitle={postTitle}
                   open={openReplies}
                   setOpen={setOpenReplies}
                 />
